+++ client/src/app/services/institutes/institutes.component.ts
@@ -47,7 +47,7 @@ import {
 } from 'app/state/curriculum/curriculum.selectors';
 import { selectAllInstitutes } from 'app/state/institute/institute.selectors';
 import { filter, Observable, Subject, take, tap } from 'rxjs';
-import { ICurriculum } from '../curriculum/curriculum.types';
+import { ICurriculum } from '../../modules/superadmin/curriculum/curriculum.types';
 import { IInstitutes } from './institutes.types';
 
 @Component({
