+++ client/src/app/state/institute/institute.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { IInstitutes } from 'app/modules/superadmin/institutes/institutes.types';
+import { IInstitutes } from 'app/services/institutes/institutes.types';
 
 export const loadInstitutes = createAction('[Institute] Load Institutes');
 export const loadInstitutesSuccess = createAction('[Institute] Load Institutes Success', props<{ institutes: IInstitutes[] }>());
