+++ client/src/app/state/institute/institute.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import * as InstituteActions from './institute.actions';
-import { IInstitutes } from 'app/modules/superadmin/institutes/institutes.types';
+import { IInstitutes } from 'app/services/institutes/institutes.types';
 
 export interface InstituteState {
   institutes: IInstitutes[];
